Revert bogus int -> size_t conversions, FATAL on 64bit platforms.
NOTE: size_t is not the same size as int, it happens to be on x86.
However, the compiler will only see "pointer to scalar, fine" and won't
complain - any dereferenced pointer access however is fatal and breaks
on SPARC64.

If flex generates "int yyleng", that's what we must use in
our pointers as well! If we want size_t, we need to have flex fixed!


git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@3460 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/msgcounts.c
@@ -23,7 +23,7 @@ AUTHOR:
 /* Globals */
 
 char	msg_count_buff[MSG_COUNT_MAX_LEN];
-size_t	msg_count_leng = MSG_COUNT_MAX_LEN;
+int	msg_count_leng = MSG_COUNT_MAX_LEN; /* DO NOT MAKE THIS SIZE_T! */
 char   *msg_count_text = msg_count_buff;
 
 const char *msg_count_header = "\"" MSG_COUNT "\"";
